# openapi/savia.openapi.yaml
openapi: 3.1.1
info:
  title: Savia Backend Executable API
  version: 0.1.0
servers:
  - url: http://127.0.0.1:3000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT verified against the configured issuer, audience, and JWKS.
    mcpOAuth:
      type: oauth2
      description: OAuth 2.1 authorization used by remote MCP clients. The MCP transport itself is not described by this OpenAPI document.
      flows:
        authorizationCode:
          authorizationUrl: https://api.savia.app/oauth/authorize
          tokenUrl: https://api.savia.app/oauth/token
          scopes:
            accounts:read: Read accounts and balances.
            transactions:read: Read financial transactions.
            transactions:write: Create and modify financial transactions.
            budgets:read: Read budgets.
            budgets:write: Modify budgets.
            reports:read: Run and read reports.
            reports:write: Create report definitions.
            workspace:admin: Administer workspace integrations.
  parameters:
    WorkspaceIdHeader:
      name: X-Workspace-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
        maxLength: 255
    IfMatch:
      name: If-Match
      in: header
      required: false
      schema: { type: string, examples: ['"7"'] }
  responses:
    Unauthorized:
      description: Unauthorized.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Forbidden:
      description: Forbidden.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    NotFound:
      description: Not found.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Conflict:
      description: Conflict.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    UnprocessableEntity:
      description: The request violates a business rule.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    PreconditionFailed:
      description: The supplied ETag does not match the current resource version.
      content:
        {
          application/problem+json:
            { schema: { $ref: '#/components/schemas/ProblemDetails' } },
        }
    ServiceUnavailable:
      description: A required service or provider is temporarily unavailable.
      content:
        {
          application/problem+json:
            { schema: { $ref: '#/components/schemas/ProblemDetails' } },
        }
  schemas:
...
    FieldViolation:
      type: object
      required:
        - field
        - code
        - message
      properties:
        field:
          type: string
        code:
          type: string
        message:
          type: string
      additionalProperties: false
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
        - code
        - traceId
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
        traceId:
          type: string
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldViolation'
          default: []
        retryable:
          type: boolean
          default: false
      additionalProperties: true
...
    DateTime:
      type: string
      format: date-time
...
    Uuid:
      type: string
      format: uuid
...
    AIProviderDescriptor:
      type: object
      required:
        [providerId, displayName, credentialTypes, enabled, policyStatus]
      properties:
        providerId: { type: string }
        displayName: { type: string }
        credentialTypes:
          {
            type: array,
            items:
              {
                type: string,
                enum:
                  [
                    api_key,
                    oauth,
                    service_account,
                    access_token,
                    gateway_token,
                    local_endpoint,
                  ],
              },
          }
        enabled: { type: boolean }
        policyStatus:
          {
            type: string,
            enum: [approved, restricted, disabled, pending_review],
          }
        models: { type: array, items: { type: string } }
      additionalProperties: false
    AICredentialMetadata:
      type: object
      required:
        [
          id,
          ownerType,
          providerId,
          credentialType,
          maskedIdentifier,
          status,
          createdAt,
        ]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        ownerType: { type: string, enum: [user, workspace] }
        providerId: { type: string }
        credentialType:
          {
            type: string,
            enum:
              [
                api_key,
                oauth,
                service_account,
                access_token,
                gateway_token,
                local_endpoint,
              ],
          }
        maskedIdentifier: { type: string }
        alias: { type: [string, 'null'] }
        status: { type: string, enum: [active, disabled, revoked] }
        lastUsedAt: { type: [string, 'null'], format: date-time }
        expiresAt: { type: [string, 'null'], format: date-time }
        createdAt: { $ref: '#/components/schemas/DateTime' }
      additionalProperties: false
    CreateAICredentialRequest:
      type: object
      required: [ownerType, providerId, credentialType, secret]
      properties:
        ownerType: { type: string, enum: [user, workspace] }
        providerId: { type: string }
        credentialType:
          {
            type: string,
            enum:
              [
                api_key,
                service_account,
                access_token,
                gateway_token,
                local_endpoint,
              ],
          }
        secret: { type: string, minLength: 1, writeOnly: true }
        alias: { type: [string, 'null'], maxLength: 120 }
        metadata: { type: object, additionalProperties: { type: string } }
      additionalProperties: false
    UpdateAICredentialRequest:
      type: object
      properties:
        alias: { type: [string, 'null'], maxLength: 120 }
        status: { type: string, enum: [active, disabled] }
        replacementSecret: { type: string, minLength: 1, writeOnly: true }
      additionalProperties: false
      minProperties: 1
    SetDefaultAIModelRequest:
      type: object
      required: [modelRef]
      properties:
        modelRef: { type: string, pattern: '^[a-z0-9][a-z0-9-]*:.+$' }
        credentialId: { type: [string, 'null'], format: uuid }
      additionalProperties: false
...
paths:
...
  /v1/ai/providers:
    get:
      operationId: listAIProviders
      summary: List supported AI providers
      tags: [AI Providers]
      responses:
        '200':
          {
            description: Provider descriptors.,
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        type: array,
                        items:
                          { $ref: '#/components/schemas/AIProviderDescriptor' },
                      },
                  },
              },
          }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
      parameters: [{ $ref: '#/components/parameters/WorkspaceIdHeader' }]
  /v1/ai/credentials:
    get:
      operationId: listAICredentials
      summary: List masked AI credential metadata
      tags: [AI Providers]
      responses:
        '200':
          {
            description: Credential metadata.,
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        type: array,
                        items:
                          { $ref: '#/components/schemas/AICredentialMetadata' },
                      },
                  },
              },
          }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
      parameters: [{ $ref: '#/components/parameters/WorkspaceIdHeader' }]
    post:
      operationId: createAICredential
      summary: Register provider-specific credential
      tags: [AI Providers]
      responses:
        '201':
          {
            description: Credential registered.,
            content:
              {
                application/json:
                  {
                    schema:
                      { $ref: '#/components/schemas/AICredentialMetadata' },
                  },
              },
          }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
      parameters:
        [
          { $ref: '#/components/parameters/WorkspaceIdHeader' },
          { $ref: '#/components/parameters/IdempotencyKey' },
        ]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema:
                    { $ref: '#/components/schemas/CreateAICredentialRequest' },
                },
            },
        }
  /v1/ai/credentials/{credentialId}:
    parameters:
      - {
          name: credentialId,
          in: path,
          required: true,
          schema: { $ref: '#/components/schemas/Uuid' },
        }
      - { $ref: '#/components/parameters/WorkspaceIdHeader' }
    patch:
      operationId: updateAICredential
      summary: Update, disable or rotate AI credential
      tags: [AI Providers]
      responses:
        '200':
          {
            description: Credential updated.,
            content:
              {
                application/json:
                  {
                    schema:
                      { $ref: '#/components/schemas/AICredentialMetadata' },
                  },
              },
          }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
      parameters:
        [
          { $ref: '#/components/parameters/IdempotencyKey' },
          { $ref: '#/components/parameters/IfMatch' },
        ]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema:
                    { $ref: '#/components/schemas/UpdateAICredentialRequest' },
                },
            },
        }
    delete:
      operationId: revokeAICredential
      summary: Revoke AI credential
      tags: [AI Providers]
      responses:
        '204': { description: Credential revoked. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
  /v1/ai/default-model:
    put:
      operationId: setDefaultAIModel
      summary: Set default provider and model
      tags: [AI Providers]
      responses:
        '204': { description: Default model updated. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
      parameters:
        [
          { $ref: '#/components/parameters/WorkspaceIdHeader' },
          { $ref: '#/components/parameters/IdempotencyKey' },
        ]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema:
                    { $ref: '#/components/schemas/SetDefaultAIModelRequest' },
                },
            },
        }
...